requirements.txt
aioconsole
aiohttp
discord.py
rgbprint
pypresence
